// src/app/risk-management/types/IBetlist.ts
export interface IBtlist {
    betTime: string
    betType: string
    consolidateId: string
    eventId: string
    eventName: string
    eventTypeId: string
    id: number
    ipAddress: string
    marketId: string
    marketName: string
    odds: number
    potLiab: number
    potProf: number
    requestedOdds: number
    selName: string
    sourceSelId: string
    sportName: string
    stake: number
    unmatched: number
    userId: string
    userName: string
  }